Create reusable hook to create fill style id
